pyproject.toml
[project]
name = "hotel-reservations"
version = "0.1.0"
description = "hotel_reservations"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "n"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
